# source/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Trackinvest Assignment",
        "description": "In this project we can view,add and update the details of Employees and Departments",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/employees/": {
            "get": {
                "summary" : "Get all the employees",
                "description" : "Get all the employees",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/employeesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/employees": {
            "post" : {
                "summary": "Save the employee details",
                "description": "Save the employee details",
                "produces":["application/json"],
                "consumes":["application/json"],
                "parameters": [
                    {
                        "in":"body",
                        "name":"body",
                        "description":"Employee object",
                        "required": true,
                        "schema": {
                                "$ref":"#/definitions/employeesResponse"
                            }
                    }
                ],
                "responses": {
                    "200": {
                        "description":"Successful operation",
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/employeesResponse"
                            }
                        }
                    }
                },
                "400": {
                    "description": "Invalid status value",
                    "schema": {
                        "$ref": "#/definitions/InvalidResponse"
                    }
                }
                
            }
        },
...
        "/department/": {
            "get": {
                "summary" : "Get all the department details",
                "description" : "Get all the department details",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/departmentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/department": {
            "post" : {
                "summary": "Save the department details",
                "description": "Save the department details",
                "produces":["application/json"],
                "consumes":["application/json"],
                "parameters": [
                    {
                        "in":"body",
                        "name":"body",
                        "description":"department object",
                        "required": true,
                        "schema": {
                                "$ref":"#/definitions/departmentResponse"
                            }
                    }
                ],
                "responses": {
                    "200": {
                        "description":"Successful operation",
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/departmentResponse"
                            }
                        }
                    }
                },
                "400": {
                    "description": "Invalid status value",
                    "schema": {
                        "$ref": "#/definitions/InvalidResponse"
                    }
                }
                
            }
        },
...
    },
    "definitions": {
        "employeesResponse": {
            "type": "object",
            "properties": {
                "empname": {
                    "type": "string"
                },
                "empdob": {
                    "type": "Date"
                },
                "deptid": {
                    "type":"integer"
                }
            }
        },
        "departmentResponse": {
            "type":"object",
            "properties": {
                "deptname": {
                    "type": "string"
                },
                "deptlocation": {
                    "type": "string"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}